+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## TEMP
 * `!tr updatelink` has been added.
 * Local hosted users persist and react command fixed.
+* CS same language now works.
+* Fixed invalid channel message.
 
 ## 1.2.6 Change Log
 * RITA is now at 6400 servers, This is just Amazing.